+++ worldoftanks/action/tankopedia_vehicles.py
@@ -57,7 +57,7 @@ class TankopediaVehiclesData:
         return clean_data
 
     def etl_data(self, application_id: str, account_id: str, token: str, load_to_db: bool, load_once: bool,
-                 realm: str) -> list:
+                 realm: str, db_path: str) -> list:
         """
         Combines all the above methods to be used as one command.
         Takes the details and the statistics data and loads it into dbsqlite.
@@ -70,11 +70,11 @@ class TankopediaVehiclesData:
         if load_to_db:
             if load_once:
                 # Checks if the data is already existing in the database else loads it.
-                if DataModelLoader.check_if_data_exists(TankopediaVehiclesModel):
+                if DataModelLoader.check_if_data_exists(TankopediaVehiclesModel, db_path=db_path):
                     logging.info('Tankopedia vehicles data will not be loaded into the database.')
                 else:
-                    DataModelLoader.insert(TankopediaVehiclesModel, clean_data)
+                    DataModelLoader.insert(TankopediaVehiclesModel, clean_data, db_path=db_path)
             else:
-                DataModelLoader.insert(TankopediaVehiclesModel, clean_data)
+                DataModelLoader.insert(TankopediaVehiclesModel, clean_data, db_path=db_path)
 
         return clean_data