+++ worldoftanks/action/vehicles_achievements.py
@@ -2,7 +2,7 @@ import logging
 
 from worldoftanks.helper.data_model_loader import DataModelLoader
 from worldoftanks.utils.api import API
-from worldoftanks.orm.data_model import VehiclesAchievements
+from worldoftanks.orm.data_model import VehiclesAchievementsModel
 
 
 class VehicleAchievementsData:
@@ -48,7 +48,8 @@ class VehicleAchievementsData:
 
         return clean_data
 
-    def etl_data(self, application_id: str, account_id: str, token: str, load_to_db: bool, realm: str) -> list:
+    def etl_data(self, application_id: str, account_id: str, token: str, load_to_db: bool, realm: str,
+                 db_path: str) -> list:
         """
         Combines all the above methods to be used as one command.
         Takes the details and the statistics data and loads it into dbsqlite.
@@ -59,6 +60,6 @@ class VehicleAchievementsData:
         clean_data = self._parse_data(raw_data=raw_data, account_id=account_id)
 
         if load_to_db:
-            DataModelLoader.insert(VehiclesAchievements, clean_data)
+            DataModelLoader.insert(VehiclesAchievementsModel, clean_data, db_path=db_path)
 
         return clean_data