+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+### [0.4.22]
+* New API feature added. Extraction of Tankopedia Badges.
+* Additional argument in the main class to specify the location of where the database should be created.
+* Updated documentation.
+* Improvements on unittest and coverage tests.
+* Removed redundant scripts.
+* Added functionality in the main class to disable loggers and loading to database.
+
 ### [Unreleased]
 ##
 - [0.3.22] Added functionality to extract data from the Player's Vehicles section.
 - [0.2.22] Added the coverage and pipeline badges to the project.
-- Removed some redundant files. 
+- Removed redundant files. 
 - [0.2.21] Added the CD / CI for testing and deployment to testing and production pypi. 
 - [0.2.2]  Added support for choosing a realm. 
 - [0.1.2]  Added more documentation on the project scope.